# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/audit-logs/export": {
      "get": {
        "description": "Streams the audit log as a downloadable PDF or XLSX file.",
        "operationId": "ExportController_audit",
        "parameters": [
          {
            "name": "format",
            "required": true,
            "in": "query",
            "description": "File format.",
            "schema": {
              "enum": [
                "pdf",
                "xlsx"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary file attachment."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Export audit logs (admin)",
        "tags": [
          "Export"
        ]
      }
    },
    "/api/inspection-reports/{id}/export": {
      "get": {
        "description": "Streams a single inspection report as PDF or XLSX. Scoped to the caller (supervisor: zone + agency).",
        "operationId": "ExportController_report",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": true,
            "in": "query",
            "description": "File format.",
            "schema": {
              "enum": [
                "pdf",
                "xlsx"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary file attachment."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Export an inspection report (supervisor/admin)",
        "tags": [
          "Export"
        ]
      }
    }
  },
  "info": {
    "title": "E-License Verification Platform API",
    "description": "REST API for the Thai e-licensing platform that plugs into the ทางรัฐ\n(Tang Rat) super app. Serves the public app and the admin portal.\n\n**Authentication.** Most endpoints require a Bearer access token\n(RS256 JWT, 15-min TTL). Obtain one from `POST /api/auth/register` or\n`POST /api/auth/login` (PUBLIC), `POST /api/auth/tang-rat`\n(PUBLIC/INSPECTOR/SUPERVISOR), or `POST /api/auth/self` (ADMIN + TOTP).\nClick **Authorize** and paste the `accessToken` to call protected routes.\nSee `docs/AUTHENTICATION.md` for a plain-English walkthrough.\n\n**Scope.** INSPECTOR/SUPERVISOR responses are filtered server-side by the\ncaller's zones and agency; clients cannot widen scope via query params.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "description": "Paste the access token returned by an /auth login route."
      }
    },
...
  }
}